+++ src/index.ts
@@ -7,6 +7,7 @@ export * from "./models/index";
 export * from "./affinity";
 export * from "./webhook-events";
 export * from "./resources/account";
+export * from "./resources/actor";
 export * from "./resources/billing";
 export * from "./resources/catalog";
 export * from "./resources/component-sessions";